+++ src/main/java/com/project/moneyj/auth/filter/JwtFilter.java
@@ -1,6 +1,8 @@
-package com.project.moneyj.auth.util;
+package com.project.moneyj.auth.filter;
 
 import com.project.moneyj.auth.dto.CustomOAuth2User;
+import com.project.moneyj.auth.util.JwtUtil;
+import com.project.moneyj.auth.util.SecurityResponseUtil;
 import com.project.moneyj.exception.MoneyjException;
 import com.project.moneyj.exception.code.AuthErrorCode;
 import com.project.moneyj.user.domain.User;
